refactor(archtest): CAPABILITY-PROVIDER-FUNNEL-01 迁 module-path-agnostic + 缩 baseline [#1639]
移 Check* 规则体到非 test capability_provider_funnel.go（platform 路径派生自
PlatformModulePath），_test.go 瘦身为 dogfood + RED fixture 单源。register=none
（gocell-internal cap_wiring allowlist，外部 vacuous）。baseline 删该行。
