+++ tools/archtest/testdata/module_path_funnel.baseline
@@ -9,7 +9,6 @@
 tools/archtest/adapter_returns_declared_types_test.go
 tools/archtest/archtest_test.go
 tools/archtest/audit_hash_input_frozen_test.go
-tools/archtest/capability_provider_funnel_test.go
 tools/archtest/celltest_import_boundary_test.go
 tools/archtest/celltest_import_scope_test.go
 tools/archtest/clock_invariants_test.go
